requirements.txt
fastapi
uvicorn
pillow
imageio
ffmpeg-python
rembg[gpu]
onnxruntime-gpu
